Iterate through declared symbols faster
